update docs + tweaks
