+++ fixtures/monorepo/base-dll/haul.config.js
@@ -5,26 +5,14 @@ const entry = withPolyfills([
   require.resolve('react'),
   require.resolve('react-native'),
   require.resolve('react-navigation'),
-  './BundleRegistry.js'
 ]);
 
 export default makeConfig({
   bundles: {
-    base_dll: {
+    base_dll: ({ dev }) => ({
       entry,
       dll: true,
-      type: 'indexed-ram-bundle',
-    },
-    // Bundle for packager server, compiled as basic-bundle (aka plain javascript)
-    // since RAM bundles are only supported when building static bundles.
-    base_dll_server: {
-      entry,
-      dll: true,
-      transform: ({ config }) => {
-        // Overwrite `output.library`, so that it will be available under `base_dll`,
-        //  instead of (by default) `base_dll_server`.
-        config.output.library = 'base_dll';
-      }
-    },
+      type: dev ? 'basic-bundle' : 'indexed-ram-bundle',
+    }),
   },
 });
