+++ fixtures/monorepo/base-dll/makeConfig.js
@@ -4,11 +4,9 @@ export default function makeConfig(copyBundle = false) {
   return ({ platform, bundleTarget, dev }) => {
     const basePath = join(
       __dirname,
-      `./dist/${platform}/${dev ? 'dev' : 'prod'}`
+      `./dist/${platform}/${dev || bundleTarget === 'server' ? 'dev' : 'prod'}`
     );
     const filename = `base_dll${
-      bundleTarget === 'server' ? '_server' : ''
-    }${
       platform === 'ios' ? '.jsbundle' : '.android.bundle'
     }`;
     return {
